+++ client/src/components/layout/Landing.js
@@ -3,76 +3,18 @@ import { Link } from "react-router-dom";
 import styles from "./styles.module.css";
 import { connect } from "react-redux";
 import PropTypes from "prop-types";
-import { getProfileByStatus } from "../../actions/profileActions";
-import SelectListGroup from "../common/SelectListGroup";
-import Spinner from "../common/Spinner";
-import ProfileItem from "../profiles/ProfileItem";
+import Profiles from "../../components/profiles/Profiles";
 
 class Landing extends Component {
   constructor(props) {
     super(props);
 
     this.state = {
-      status: "",
-      errors: {}
+      status: ""
     };
-
-    this.onChange = this.onChange.bind(this);
-    this.onSubmit = this.onSubmit.bind(this);
-  }
-
-  componentDidMount() {
-    this.props.getProfileByStatus();
-  }
-
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.errors) {
-      this.setState({ errors: nextProps.errors });
-    }
-  }
-
-  onSubmit(e) {
-    e.preventDefault();
-
-    /*const status = {
-      status: this.state.status
-    };*/
-
-    this.props.getProfileByStatus(this.state.status);
-  }
-
-  onChange(e) {
-    this.setState({ [e.target.name]: e.target.value });
   }
 
   render() {
-    const { errors } = this.state;
-    const { profiles, loading } = this.props.profile;
-    let profileItems;
-
-    if (profiles === null || loading) {
-      profileItems = <Spinner />;
-    } else {
-      if (profiles.length > 0) {
-        profileItems = profiles.map(profile => (
-          <ProfileItem key={profile._id} profile={profile} />
-        ));
-      } else {
-        profileItems = <h4>No profiles found...</h4>;
-      }
-    }
-
-    const options = [
-      { label: "* Select Professional Status", value: 0 },
-      { label: "Developer", value: "Developer" },
-      { label: "Junior Developer", value: "Junior Developer" },
-      { label: "Senior Developer", value: "Senior Developer" },
-      { label: "Manager", value: "Manager" },
-      { label: "Student or Learning", value: "Student or Learning" },
-      { label: "Instructor or Teacher", value: "Instructor or Teacher" },
-      { label: "Intern", value: "Intern" },
-      { label: "Other", value: "Other" }
-    ];
     return (
       <div className="landing">
         <div className="dark-overlay landing-inner text-light">
@@ -92,23 +34,7 @@ class Landing extends Component {
                 <Link to="/login" className={styles.NavLink}>
                   Login
                 </Link>
-                <form onSubmit={this.onSubmit}>
-                  <SelectListGroup
-                    placeholder="Status"
-                    name="status"
-                    value={this.state.status}
-                    onChange={this.onChange}
-                    options={options}
-                    error={errors.status}
-                    info="Give us an idea of where you are at in your career"
-                  />
-                  <input
-                    type="submit"
-                    value="Submit"
-                    className="btn btn-info btn-block mt-4"
-                  />
-                  <div>{profileItems}</div>
-                </form>
+                <Profiles status={this.state.status} />
               </div>
             </div>
           </div>
@@ -119,7 +45,6 @@ class Landing extends Component {
 }
 
 Landing.propTypes = {
-  getProfileByStatus: PropTypes.func.isRequired,
   profile: PropTypes.object.isRequired
 };
 
@@ -127,7 +52,4 @@ const mapStateToProps = state => ({
   profile: state.profile
 });
 
-export default connect(
-  mapStateToProps,
-  { getProfileByStatus }
-)(Landing);
+export default connect(mapStateToProps)(Landing);
