+++ client/src/components/profiles/Profiles.js
@@ -1,29 +1,89 @@
 import React, { Component } from "react";
+import { Link } from "react-router-dom";
 import { connect } from "react-redux";
 import PropTypes from "prop-types";
 import Spinner from "../common/Spinner";
 import { getProfiles } from "../../actions/profileActions";
 import ProfileItem from "./ProfileItem";
+import { getProfileByStatus } from "../../actions/profileActions";
+import SelectListGroup from "../common/SelectListGroup";
+import isEmpty from "../../validation/is-empty";
+import { withRouter } from "react-router-dom";
 
 class Profiles extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      status: "",
+      filteredByStatus: [],
+      errors: {}
+    };
+
+    this.onChange = this.onChange.bind(this);
+    this.onSubmit = this.onSubmit.bind(this);
+  }
+
   componentDidMount() {
     this.props.getProfiles();
   }
+
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.errors) {
+      this.setState({ errors: nextProps.errors });
+    }
+    if (nextProps.profile.profile) {
+      this.setState({ filteredByStatus: nextProps.profile.profile });
+    }
+  }
+
+  onSubmit(e) {
+    e.preventDefault();
+
+    this.props.getProfileByStatus(this.state.status, this.props.history);
+  }
+
+  onChange(e) {
+    this.setState({ [e.target.name]: e.target.value });
+  }
+
   render() {
     const { profiles, loading } = this.props.profile;
     let profileItems;
 
-    if (profiles === null || loading) {
-      profileItems = <Spinner />;
+    const { errors } = this.state;
+    const { filteredByStatus } = this.state;
+    let profileStatusItems;
+
+    console.log(profiles);
+
+    if (isEmpty(profiles) || loading) {
+      profileStatusItems = <Spinner />;
     } else {
-      if (profiles.length > 0) {
+      if (!isEmpty(profiles) && isEmpty(filteredByStatus)) {
         profileItems = profiles.map(profile => (
+          <ProfileItem key={profile._id} profile={profile} />
+        ));
+      } else if (filteredByStatus.length > 0) {
+        profileStatusItems = filteredByStatus.map(profile => (
           <ProfileItem key={profile._id} profile={profile} />
         ));
       } else {
         profileItems = <h4>No profiles found...</h4>;
       }
     }
+
+    const options = [
+      { label: "* Select Professional Status", value: 0 },
+      { label: "Developer", value: "Developer" },
+      { label: "Junior Developer", value: "Junior Developer" },
+      { label: "Senior Developer", value: "Senior Developer" },
+      { label: "Manager", value: "Manager" },
+      { label: "Student or Learning", value: "Student or Learning" },
+      { label: "Instructor or Teacher", value: "Instructor or Teacher" },
+      { label: "Intern", value: "Intern" },
+      { label: "Other", value: "Other" }
+    ];
 
     return (
       <div className="profiles">
@@ -34,6 +94,24 @@ class Profiles extends Component {
               <p className="lead text-center">
                 Browse and connect with developers
               </p>
+              <form onSubmit={this.onSubmit}>
+                <SelectListGroup
+                  placeholder="Status"
+                  name="status"
+                  value={this.state.status}
+                  onChange={this.onChange}
+                  options={options}
+                  error={errors.status}
+                  info="Give us an idea of where you are at in your career"
+                />
+                <input
+                  type="submit"
+                  value="Submit"
+                  className="btn btn-info btn-block mt-4"
+                />
+
+                <div>{profileStatusItems}</div>
+              </form>
               {profileItems}
             </div>
           </div>
@@ -45,6 +123,7 @@ class Profiles extends Component {
 
 Profiles.propTypes = {
   getProfiles: PropTypes.func.isRequired,
+  getProfileByStatus: PropTypes.func.isRequired,
   profile: PropTypes.object.isRequired
 };
 
@@ -54,5 +133,5 @@ const mapStateToProps = state => ({
 
 export default connect(
   mapStateToProps,
-  { getProfiles }
-)(Profiles);
+  { getProfiles, getProfileByStatus }
+)(withRouter(Profiles));
